+++ packages/common/ProForm/src/types/formCustomizerTypes.ts
@@ -74,6 +74,7 @@ export interface ItemSchema {
 	show?: boolean;
 	required?: boolean;
 	placeholder?: string;
+	runtime?: Runtime;
 	native?: NativeCustomizationOptions;
 	grid?: GridStyle;
 	label?: string;